[EI-227] [web] refactor: Refactor indicator board component
